# swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/solutions": {
      "get": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Get available solutions list.",
        "operationId": "GetSolutionsList",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "available solutions",
            "schema": {
              "$ref": "#/definitions/AvailableSolutionsList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Add available solution.",
        "operationId": "AddTemplate",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AvailableSolution"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "solution added"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/solutions/{solution}": {
      "put": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Update available solution.",
        "operationId": "UpdateTemplate",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AvailableSolution"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "solution updated"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Delete available solution.",
        "operationId": "DeleteTemplate",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "solution deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/solutions/{solution}/activate": {
      "post": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Update available solution.",
        "operationId": "ActivateTemplate",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "solution activated"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/solutions/{solution}/deactivate": {
      "post": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Update available solution.",
        "operationId": "DeactivateTemplate",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "solution deactivated"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/solutions/{solution}/env": {
      "get": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Get available solution environment variables.",
        "operationId": "GetTemplatesEnv",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "available solution envs",
            "schema": {
              "$ref": "#/definitions/SolutionEnv"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/solutions/{solution}/resources": {
      "get": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Get available solution resources.",
        "operationId": "GetTemplatesResources",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "available solution resources",
            "schema": {
              "$ref": "#/definitions/SolutionResources"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/user_solutions": {
      "get": {
        "tags": [
          "UserSolutions"
        ],
        "summary": "Get running solutions list.",
        "operationId": "GetSolutionsList",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "running solutions list",
            "schema": {
              "$ref": "#/definitions/UserSolutionsList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "UserSolutions"
        ],
        "summary": "Run solution.",
        "operationId": "RunSolution",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserSolution"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "solution created",
            "schema": {
              "$ref": "#/definitions/RunSolutionResponce"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
  },
  "definitions": {
    "AvailableSolution": {
      "description": "AvailableSolution -- solution which user can run",
      "type": "object",
      "properties": {
        "Active": {
          "type": "boolean"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Images"
        },
        "limits": {
          "$ref": "#/definitions/Limits"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "url": {
          "type": "string",
          "x-go-name": "URL"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
    "AvailableSolutionsList": {
      "description": "AvailableSolutionsList -- list of available solutions",
      "type": "object",
      "properties": {
        "solutions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AvailableSolution"
          },
          "x-go-name": "Solutions"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
...
    "Limits": {
      "description": "Limits -- solution resources limits",
      "type": "object",
      "properties": {
        "cpu": {
          "type": "string",
          "x-go-name": "CPU"
        },
        "ram": {
          "type": "string",
          "x-go-name": "RAM"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
    "RunSolutionResponce": {
      "description": "RunSolutionResponce -- responce to run solution request",
      "type": "object",
      "properties": {
        "created": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Created"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Errors"
        },
        "not_created": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "NotCreated"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
    "SolutionEnv": {
      "description": "SolutionEnv -- solution environment variables",
      "type": "object",
      "properties": {
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Env"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
    "SolutionResources": {
      "description": "SolutionResources -- list of solution resources",
      "type": "object",
      "properties": {
        "resources": {
          "type": "object",
          "additionalProperties": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "Resources"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
    "UserSolution": {
      "description": "UserSolution -- running solution",
      "type": "object",
      "required": [
        "template",
        "name",
        "namespace"
      ],
      "properties": {
        "branch": {
          "type": "string",
          "x-go-name": "Branch"
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Env"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "template": {
          "type": "string",
          "x-go-name": "Template"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
    "UserSolutionsList": {
      "description": "UserSolutionsList -- list of running solution",
      "type": "object",
      "properties": {
        "solutions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserSolution"
          },
          "x-go-name": "Solutions"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
  },
  "parameters": {
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}